derive preset ranges on read instead of caching them
Cached bounds went stale in place. updateDatesFromRange ran only from the
constructor and handleRangeClick, so endDate was frozen at whatever
new Date() was when the range was last chosen, and filterRows is
inclusive against it — a session logged after the click was fetched by
revalidate() and then filtered straight back out. An installed PWA
resumed the next morning showed "1M as of yesterday" until the user
clicked a range button. Same class of staleness this branch already
fixed at the start of the window, at the other end.

getRange() now calls presetBounds() per read for every range but CUSTOM,
which keeps its explicit pair. The stored pair survives only to prefill
the Custom inputs.

The arithmetic moves to src/dateRange.js, which imports nothing: the
caller injects getBegin (it throws until setBegin has seen data, and App
constructs the widget before that). That keeps the calendar tests off d3
and the gitignored alias table, and makes the range switch itself
testable — it is where both clamping bugs actually lived, and no test
covered it. The widget is loaded by one test file now, for the wiring
that only it can show.
