+++ CLAUDE.md
@@ -98,7 +98,8 @@ Hash routing lives in `NavigationComponent`: menu clicks set `window.location.ha
 
 **UI:**
 - `NavigationComponent` (`src/navigationComponent.js`) — hamburger menu (native dismiss: outside-click + Escape), segmented Part buttons (V1/V2/VA/ANY), Player multiselect dropdown, view switching + hash routing. Delegates the date range to `DateFilterWidget`.
-- `DateFilterWidget` (`src/dateFilterWidget.js`) — reusable segmented date range picker (`All` / `YTD` / `1Y` / `6M` / `1M` / `Custom`). Class-based selectors scoped to mount point so multiple instances can coexist; Home and Dashboard each have their own, so a new range option appears on both for free. The relative ranges (`1Y`, `6M`, `1M`) start at the same day of the month N months back via the pure exported `monthsAgo`, which clamps to that month's last day when it is shorter (Mar 31 → Feb 29/28, and Feb 29 → Feb 28 a year back) instead of letting `setMonth`/`setFullYear` overflow forward into the next month. Every preset range start is then anchored to local midnight via `startOfDay`, so a window is a whole number of days and a session logged earlier on the boundary day still counts — matching the Custom inputs, which anchor at midnight through `fromDateInputValue`.
+- `DateFilterWidget` (`src/dateFilterWidget.js`) — reusable segmented date range picker (`All` / `YTD` / `1Y` / `6M` / `1M` / `Custom`). Class-based selectors scoped to mount point so multiple instances can coexist; Home and Dashboard each have their own, so a new range option appears on both for free. The class is chrome plus the `Custom` pair; the arithmetic is `src/dateRange.js`.
+- `dateRange` (`src/dateRange.js`) — the range math, pure and import-free (the caller injects `getBegin`, which throws until `setBegin` has seen data — and `App` constructs `NavigationComponent`, hence the widget, before that happens). `presetBounds(rangeId, now, getBeginDate)` maps a range id to its `[start, end]`; `getRange()` calls it **on every read** for every range but `Custom`, so a relative window means "the last month" rather than "the month before the button was pressed". Caching that pair is a live bug, not a style question: `filterEngine.filterRows` is inclusive against the end, so a session logged after the click gets fetched by `revalidate()` and filtered straight back out, and a PWA resumed the next morning shows "1M as of yesterday" until a range button is clicked. The relative ranges (`1Y`, `6M`, `1M`) start at the same day of the month N months back via `monthsAgo`, which clamps to that month's last day when it is shorter (Mar 31 → Feb 29/28, and Feb 29 → Feb 28 a year back) instead of letting `setMonth`/`setFullYear` overflow forward into the next month. Every start is then anchored to local midnight by `startOfDay`, so a window is a whole number of days and a session logged earlier on the boundary day still counts — matching the Custom inputs, which anchor at midnight through `fromDateInputValue`. `ALL` is the one branch where that anchor is a no-op today, since `setBegin` already normalizes to the 1st of the month at midnight.
 - `TabComponent` (`src/tabComponent.js`) — per-composer tab content + ALL tab. `updateTabContent` early-returns to `updateAllTabContent` for the special ALL tab. Random-button suggestion respects current filters (pure `pickRandomWork`, rebound on every update); grouping is the pure `groupPlaysByWork`. `activeTab` on the instance is the source of truth (`.active-tab` classes are reflections); `onTabShown` hook drives lazy rendering (below).
 - `CalendarComponent` (`src/calendarComponent.js`) — calendar grid, legend, per-year stats column, "Last 365 days" header (uses `renderRecentStats` → `computeAggregateStats`). Legend SVG and grid are width-coupled via CSS.
 - `MusicianNetworkComponent` (`src/musicianNetworkComponent.js`) — state machine + chrome (tabs, slider via pure `computeSliderSync`, fullscreen, name toggle, tooltip builders); the three views render via `networkGraphRenderer.js` / `networkMatrixRenderer.js` / `networkChordRenderer.js`, each taking a plain ctx object.
